+++ CPU_Vx/Assembler/Printers/BinaryRamPrinter.h
@@ -3,6 +3,7 @@
 #include <iostream>
 #include <vector>
 #include <fstream>
+#include <cstdint>
 
 
 class BinaryRamPrinter
@@ -11,7 +12,7 @@ public:
 	BinaryRamPrinter();
 	~BinaryRamPrinter();
 
-	void run(std::string path, std::vector<int>& binaryData);
+	void run(std::string path, std::vector<uint8_t>& binaryData);
 private:
 
 };
